refactor: Move toward separating pumpkin into separate crates (#218)
We start moving toward a design where different features of the solver
are part of different crates. In the long term, this should make the
library more useful as not just the solver is the public API, but also
parts of its internals.

This is a first step toward that idea, where we create a `core` crate
that is separate from the `pumpkin_solver` crate. Over the course of
future PRs, different components can be extracted.