feat: Add objective brancher to default search (#230)
In some cases, the objective variable is not fixed for some models (e.g.
`sparse_mds`); in this case, it will use the backup selector which will
randomly pick a variable and randomly assign it.

However, we can attempt to fix it to its extremum (depending on whether
it is minimising or maximising).