fix: Reporting conflict if there is a task which uses more resource than capacity (#236)
Currently, we do not proactively look for the case where a task is
provided that uses more resource than the capacity; this leads to an
issue when this conflict is reported as it could have been reported
earlier.

This PR addresses this issue by performing an additional check when
initialising. This occurs in [certain
models](https://github.com/MiniZinc/mzn-challenge/blob/develop/2024/aircraft-disassembly/aircraft.mzn).