Three-tier nogood management system (#213)
I drafted a new nogood management strategy. The previous version was a
"temporary" solution that stayed for too long; this version is supposed
to be better.

There are a few changes:
- This version implements the three-tier system for nogoods, this is the
main change. This means that nogoods are partitioned into three
categories (low-, mid-, and high-lbd), and the solver keeps them around
for a different amount of time. Low-lbd are typically kept longer (most
likely forever), high-lbd tier is temporary (similar to what we had
before), whereas now there is also a mid-tier with lbds in between high
and low, which is the main change compared to before.
- The values for low-lbd is now 3, and high-lbd is 7 or greater. We
might need to try out different values to see if there is any difference
for CP benchmarks.
- We now also delete _some_ nogoods falsified at the root level. When
cleaning up nogoods, if we happen to encounter a nogood falsified at the
root level (one of its watchers is falsified at the root, or its cached
predicate is falsified), then we will also remove that nogood. It could
very well be the case that some nogoods are falsified, but we do not
detect them in this pass (if we encounter that nogood later, its
watchers will be updated so it will get removed later). We also do not
simplify nogoods by removing true predicates, which could come later.

Can you please have a detailed look at the code, to see if something
looks strange, e.g., does the code follow the comments?

In particular, there is an issue that surprises me, and where the tests
fail. Have a look at the function
`are_watched_predicates_falsified_at_root_level`. For some reason, when
running the tests, getting a decision level for a predicate that is
assigned fails - how is that possible?

---------

Co-authored-by: Emir Demirović <edemirovic@tudelft.nl>
Co-authored-by: Imko Marijnissen <imko.marijnissen@gmail.com>