feat: Remove unnecessary variables introduced by bool2int (#232)
Currently, when we encounter a constraint `bool2int(a, b)`, we create a
new variable `a` and a new variable `b`; this PR aims to address this
issue by simply using the DomainId of `b` in place of `a`.

To achieve this, we removed the distinction of boolean and integer variables from the FlatZinc compiler. Now an equivalence class can contain variables of both those types.

- [x] Run using MiniZinc tests

---------

Co-authored-by: Maarten Flippo <maartenflippo@outlook.com>