feat(pumpkin-solver-py): Setup publishing of the python interface (#277)
This PR sets up the python interface for publishing to PyPi. It includes
a workflow that builds the library for all platforms and uploads them to
the package repository. For now that workflow is triggered manually, but
in future it should be integrated into the publishing of the rest of the
solver's packages.