refactor: Minimising and lifting the explanations for empty domains (#283)
Currently, the explanations for empty domain explanations contain the
explanation of the last-propagated predicate and the bounds at the time
of that propagation of the domain which became empty.

However, based on the predicate which was propagated, it is possible to
simplify the explanation. In all cases, this reduces the number of added
predicates from two to one while also allowing the possibility of
lifting the explanation.

On certain scheduling instances, this improved the number of nodes
reached to find the same solution and it decreased the time to this same
solution.

## TODO
- [x] Profile on a larger benchmark set

## MiniZinc Score
- Before Change: 172.42
- After Change: 187.58

## Primal Integral
- Before Change: 34.49
- After Change: 32.92