feat: Allow unbounded integers (#285)
We saw in the MiniZinc challenge that we were not able to solve
instances where integers were unbounded; this PR aims to address this by
creating "unbounded" integers as regular variables with some predefined
bounds.

The original idea was to make it `[i32::MIN, i32::MAX]`, but this led to
overflows in affine views on these instances. For now, I have taken the
bounds which are used by Chuffed, but any comments about this would be
welcome!

_Some small adjustments to the random splitter since it does not work
well with extreme bounds at the moment_