feat: Avoid unnecessary traversals of profiles in time-tabling + Improving Explanations (#282)
This PR introduces a number of changes to the time-table propagator(s):
- One of the main differences is that the propagation will use the
method
[Vec::partition_point(..)](https://doc.rust-lang.org/std/vec/struct.Vec.html#method.partition_point)
to avoid looking at profiles which are guaranteed not to be able to
update the profiles
- The method which generated the sequence of profiles was looking for
_consecutive_ profiles which caused the maximum (minimum) lower-bound
(upper-bound) update. However, imagine the scenario in which there are 3
profiles where the 1st and the 3rd are able to (together) propagate the
lower-bound to its maximum value but the 2nd profile cannot propagate
the task; the current implementation would only generate an explanation
based on the 1st profile and then keep looking for updates until it
finds the 3rd profile. This can be especially detrimental for pointwise
explanations. Thus, this PR addresses this by allowing a non-consecutive
sequence of profiles for propagations.
- Finally, the defaults are changed such that the default explanation is
the pointwise explanation, and by default we go over each task and check
whether it can be updated, allowing the previous optimisations.