feat(pumpkin-solver): Adding edge-finding for the disjunctive constraint (#275)
Adds my implementation of edge-finding for the disjunctive constraint
based on `Computing Explanations for the Unary
Resource Constraint - Petr Vilím` and `Evaluating the Impact of
Explanations on the Performance of an Edge-Finding Propagator - Radu
Andrei Vasile`.

- [x] Perform evaluation

I have performed the evaluation on [this set of 75 jobshop
benchmarks](https://github.com/Bradul/Pumpkin/tree/feat/scheduling-project/Benchmarks/OR_tests)
(while using the `smallest` and `indomain_min` strategy for search).

## Baseline
```
{'OPTIMAL': 9, 'SATISFIABLE': 66}
```
**MiniZinc Score**: 29.26

## Edge-finding
```
{'OPTIMAL': 24, 'SATISFIABLE': 51}
```
**MiniZinc Score**: 45.74

## Overall
In 16 instances, the two approaches find the same solution. In 52
instances, using edge-finding results in a better solution. In 7
instances, the baseline results in a better solution.

Overall, it seems that edge-finding performs much better than the
decomposition when the problem is focussed on `disjunctive` constraints.