feat(pumpkin-solver): Lazy explanation for linear less than or equals (#290)
Currently, we use an eager explanation for `<=` constraints which can be
quite expensive if there are many of them, each making several
allocations.

This PR addresses this by lazily calculating the reason for these
propagations. Additionally, in my mind it makes sense to include the
predicate for which the reason is asked when asking for a lazy reason.

_Could also be related to #284; perhaps only doing the trimming for the
lazy explanations is worthwhile_

## TODO
- [x] Profiling
- [x] Merge #297 

**These results are _with_ the changes from #297; without these changes,
the MiniZinc score is worse**

### MiniZinc Scoring
- Main: 173.89
- **Lazy Explanations: 186.12**

### Primal Integral
- **Main: 26.13**
- Lazy Explanations: 26.69