refactor(pumpkin-core): Use a typed propagator handle to identify propagators (#302)
Propagator IDs are type-erased handles to propagator instances. However,
throughout the solver we rely on the nogood propagator in particular to
be downcastable. In this PR we introduce an API that makes it easy to
obtain concrete propagator types from the propagator store.
Additionally, we use this API in all places that previously relied on a
hard-coded ID for the nogood propagator.