refactor(pumpkin-core): Separate conflict resolvers further from solver (#310)
The previous design of the `ConflictResolver` trait assumed learning of
nogoods. We are now moving to using pumpkin to learn other constraint
types as well. For that, we need a clearer boundary of what a conflict
resolver is, without assuming we will learn nogoods.

In this first change, we move _all_ the conflict resolution that is
currently implemented to the `ResolutionResolver`. The solver itself
only issues a call to `ConflictResolver::resolve_conflict` and can
proceed with search after that function exits.

To fully decouple the learning from the solver, we need one more change.
The conflict info for empty domains should not automatically be a
nogood. Instead, it should be the information required to compute the
conflict nogood for the empty domain case. The `ResolutionResolver`
itself then has to compute the conflict nogood. Note that this is
already in the documentation of `ConflictResolver`.

This change is not made in this PR. The scope is already significant, so
to aid in reviewing I did not want to make it any bigger.