feat: Introducing `fzn-rs`, a better FlatZinc parser that is easier to re-use across projects (#238)
Our current FlatZinc parser implementation works, but it has a few
flaws:
1. It is difficult to support both the original `fzn` format, as well as
the new JSON-based format.
2. The error messages have very little information. They lack any
indication of where the error happened, which can make for difficult
debugging.
3. We write a lot of code to resolve variables, constants, etc. All that
code has to be repeated when implementing other tools that consume
FlatZinc (e.g. a proof processor or proof checker).

That is why this PR introduces a new FlatZinc parsing crate, which
provides the following:
- A new AST that is heavily inspired by
[flatzinc-serde](https://docs.rs/flatzinc-serde). It improves the AST we
obtain from [flatzinc](https://docs.rs/flatzinc) mainly because we do
not attempt to infer type information of identifiers in constraint
arguments.
- The AST tracks, for every node, where it originated in the source.
This allows for more descriptive error messages.
- Derive macros that allow for easy specification of the shape of
constraints that are supported by the consumer of the crate. No more
resolving constants/variables/arrays. That is all done in generated
code. Look at the crate documentation for `fzn-rs` for an idea of what
that looks like.