fix(pumpkin-core): Allow propagators to register for predicates becoming true (#331)
Predicate notification only happened for the `NogoodPropagator`. Now
that we are moving toward a more flexible API, we need arbitrary
propagators to be able to register for predicates becoming true.

This patch does _not_ remove the special status of the
`NogoodPropagator`, it merely enables other propagators to register for
specific events.