Fix sign error in greater_than helper
`greater_than(terms, rhs)` is implemented as
`greater_than_or_equals(terms, rhs - 1)`, but should be
`greater_than_or_equals(terms, rhs - 1)`. This is likely a copy-paste
error from the implementation of `less_than`.

Note: I've added a unit test saying that "x > 0" is not satisfiable for
an integer variable x which is required to be 0. This feels like an odd
test to have. I'm happy to remove or change it.