fix: Generate constraint tags via State (#335)
Currently, we can only generate `ConstraintTag`s via the solver;
however, all propagators take as input a constraint tag. This means that
we have an odd dependency between having to create a `Solver` (which
contains a `State`) to create a `ConstraintTag` before we can post a
propagator to the `State`.

This PR addresses this by moving the generation of `ConstraintTag`s via
the `State`.