feat(pumpkin-core): Crash if a propagator constructor does not register anything (#338)
If a `PropagatorConstructor` implementation does not register the
propagator to be enqueued for any events, the propagator will never be
enqueued except for when it is added to the `State`. In most cases, this
is a programming error where someone forgets to call register.

To make it more difficult for this error to occur, the propagator
constructor context will now crash if it is dropped without any
registration having taken place. In the odd case where no registration
is done intentionally,
`PropagatorConstructorContext::will_not_register_for_any_events` can be
called to avoid the crash.